pyproject.toml
[project]
name = "x-monitor"
version = "1.0.0"
description = "A tool to monitor X user activities using Playwright and uv."
requires-python = ">=3.8"
dependencies = [
    "playwright",
    "playwright-stealth>=1.0.6",
    "python-dotenv",
    "requests",
]

[tool.uv]
dev-dependencies = []
